Isolate SplitRule builder in trait
